feat: add checking permission

+++ pkl/application/modules/master_data/controllers/Lecturer.php
@@ -6,6 +6,14 @@ class Lecturer extends MY_Controller
     public function __construct()
     {
         parent::__construct();
+
+        if (!$this->check_permission('master_data', 'read')) {
+            echo 'Access Denied!!!';
+            $this->output->set_status_header(403);
+            $this->output->set_output('Access Denied');
+            exit;
+        }
+
         $this->load->model('Lecturer_model');
     }
 
@@ -22,6 +30,12 @@ class Lecturer extends MY_Controller
             show_error('Access Denied', 403);
         }
 
+        if (!$this->check_permission('master_data', 'create')) {
+            $this->output->set_content_type('application/json');
+            $this->output->set_output(json_encode(['status' => 'access_denied']));
+            return;
+        }
+
         $data['roles'] = $this->Lecturer_model->get_all_roles();
         $data['title'] = 'Dosen';
 
@@ -34,6 +48,12 @@ class Lecturer extends MY_Controller
     {
         if (!$this->input->is_ajax_request()) {
             show_error('Access Denied', 403);
+        }
+
+        if (!$this->check_permission('master_data', 'create')) {
+            $this->output->set_content_type('application/json');
+            $this->output->set_output(json_encode(['status' => 'access_denied']));
+            return;
         }
 
         $result = $this->Lecturer_model->create() && $this->Lecturer_model->create_roles() ? 'success' : 'error';
@@ -61,6 +81,12 @@ class Lecturer extends MY_Controller
             show_error('Access Denied', 403);
         }
 
+        if (!$this->check_permission('master_data', 'update')) {
+            $this->output->set_content_type('application/json');
+            $this->output->set_output(json_encode(['status' => 'access_denied']));
+            return;
+        }
+
         $id = $this->input->post('id', true);
         $data = $this->Lecturer_model->get_data_detail($id);
         $data['roles'] = $this->Lecturer_model->get_all_roles();
@@ -76,6 +102,12 @@ class Lecturer extends MY_Controller
             show_error('Access Denied', 403);
         }
 
+        if (!$this->check_permission('master_data', 'update')) {
+            $this->output->set_content_type('application/json');
+            $this->output->set_output(json_encode(['status' => 'access_denied']));
+            return;
+        }
+
         $result = $this->Lecturer_model->update() && $this->Lecturer_model->create_roles() ? 'success' : 'error';
         $this->output->set_content_type('application/json')->set_output(json_encode(['status' => $result]));
     }
@@ -84,6 +116,12 @@ class Lecturer extends MY_Controller
     {
         if (!$this->input->is_ajax_request()) {
             show_error('Access Denied', 403);
+        }
+
+        if (!$this->check_permission('master_data', 'delete')) {
+            $this->output->set_content_type('application/json');
+            $this->output->set_output(json_encode(['status' => 'access_denied']));
+            return;
         }
 
         $result = $this->Lecturer_model->delete() ? 'success' : 'error';
